# Speed up boot time by installing bootnstrap