--- a/Gemfile
+++ b/Gemfile
@@ -67,6 +67,7 @@ gem 'turnout'
 gem 'twilio-ruby'
 gem 'uglifier'
 gem 'webpacker', git: 'https://github.com/rails/webpacker.git'
+gem 'bootsnap', require: false
 
 group :development, :test do
   gem 'byebug'