https://github.com/Shopify/bootsnap

Speeds up rails boot time by 50% per README. My test was even faster:

```
# Before
➜ rod:~/code/sou/champaign(development)% time rails r "puts 'helo'"
helo
rails r "puts 'helo'"  5.02s user 1.98s system 94% cpu 7.369 total
# After
➜ rod:~/code/sou/champaign(development) ✗% time rails r "puts 'helo'"
helo
rails r "puts 'helo'"  2.78s user 0.50s system 120% cpu 2.714 total
```